+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow transformers to work on top-level objects
 - Separate transformer to merge schemas containing `allOf`
 
+### Changed
+- Add separate types for `ProductSchema`
+
 ## [0.8.0]
 ### Added
 - Allow merging schemas by intersection instead of union